store to map
